# feat: more performant calculation of derived resources