--- a/src/bartiq/compilation/_compile.py
+++ b/src/bartiq/compilation/_compile.py
@@ -388,11 +388,36 @@ def _compile(
         repetition=repetition,
         children_order=routine.children_order,
     )
-    return _add_derived_resources(compiled_routine, backend, derived_resources)
+
+    tmp_routine = (
+        compiled_routine
+        if CompilationFlags.EXPAND_RESOURCES in compilation_flags
+        else _dummify_child_resources(compiled_routine, backend)
+    )
+    tmp_routine = _add_derived_resources(tmp_routine, backend, derived_resources)
+
+    return replace(compiled_routine, resources=tmp_routine.resources)
 
 
 def _accepts_resource_name(func: Calculate[T] | CalculateWithName[T]) -> TypeIs[CalculateWithName[T]]:
     return "resource_name" in inspect.signature(func).parameters
+
+
+def _dummify_resources(compiled_routine: CompiledRoutine[T], backend: SymbolicBackend[T]) -> CompiledRoutine[T]:
+    return replace(
+        compiled_routine,
+        resources={
+            name: replace(res, value=backend.as_expression(f"{compiled_routine.name}.{name}"))
+            for name, res in compiled_routine.resources.items()
+        },
+    )
+
+
+def _dummify_child_resources(compiled_routine: CompiledRoutine[T], backend: SymbolicBackend[T]) -> CompiledRoutine[T]:
+    return replace(
+        compiled_routine,
+        children={cname: _dummify_resources(child, backend) for cname, child in compiled_routine.children.items()},
+    )
 
 
 def _add_derived_resources(