--- a/src/bartiq/compilation/_compile.py
+++ b/src/bartiq/compilation/_compile.py
@@ -392,7 +392,7 @@ def _compile(
     tmp_routine = (
         compiled_routine
         if CompilationFlags.EXPAND_RESOURCES in compilation_flags
-        else _dummify_child_resources(compiled_routine, backend)
+        else _introduce_placeholder_child_resources(compiled_routine, backend)
     )
     tmp_routine = _add_derived_resources(tmp_routine, backend, derived_resources)
 
@@ -403,7 +403,9 @@ def _accepts_resource_name(func: Calculate[T] | CalculateWithName[T]) -> TypeIs[
     return "resource_name" in inspect.signature(func).parameters
 
 
-def _dummify_resources(compiled_routine: CompiledRoutine[T], backend: SymbolicBackend[T]) -> CompiledRoutine[T]:
+def _introduce_placeholder_resources(
+    compiled_routine: CompiledRoutine[T], backend: SymbolicBackend[T]
+) -> CompiledRoutine[T]:
     return replace(
         compiled_routine,
         resources={
@@ -413,10 +415,15 @@ def _dummify_resources(compiled_routine: CompiledRoutine[T], backend: SymbolicBa
     )
 
 
-def _dummify_child_resources(compiled_routine: CompiledRoutine[T], backend: SymbolicBackend[T]) -> CompiledRoutine[T]:
+def _introduce_placeholder_child_resources(
+    compiled_routine: CompiledRoutine[T], backend: SymbolicBackend[T]
+) -> CompiledRoutine[T]:
     return replace(
         compiled_routine,
-        children={cname: _dummify_resources(child, backend) for cname, child in compiled_routine.children.items()},
+        children={
+            cname: _introduce_placeholder_resources(child, backend)
+            for cname, child in compiled_routine.children.items()
+        },
     )