## Description

This PR introduced performance improvements for calculating derived resources, as well as specific improvements to the qubit highwater calculation.

So far, the process of calculating derived resources wasn't respecting the compilation flags. Therefore, even if other resources weren't being expanded for usual resources, they were being expanded for the purpose of computing the derived ones. This was undermining our previous optimization efforts, as the highwater (which is a derived resource) calculation is currently the biggest bottleneck. This PR solves this problem by temporarily replacing child resources with dummy variables, unless EXPAND_RESOURCES flag is passed. Such "dummified" routine is passed to the resource-computing functions, which can access resource values per usual - which makes this change fully backwards compatible.

Additionally, profiling the double factorization example (the new one, not the one in one of our test yamls), revealed that a lot of time is spend spinning wheels in the `sympy.Max` function. This is because sympy makes a significant effort trying to simplify `Max` but is ultimately unable to. This is unsurprising, given the complexity of the expressions occuring in port sizes and the fact that we don't have assumptions system. Therefore, I replaced the default `Max` with a custom one, which puts significantly less effort simplifying itself.

As a result of both the described changes, there is a following improvement in running new DF example (this is a representable result, not an average, but the results don't vary that much between runs):

**Before changes**:

- Compilation time:  137.3914284160128s
- Evaluation_time:  482.83837754197884s

**After changes**:

- Compilation time:  4.762517333030701s
- Evaluation_time:  2.800215625029523s


## Please verify that you have completed the following steps

- [x] I have self-reviewed my code.
- [x] I have included test cases validating introduced feature/fix.
- [ ] I have updated documentation.